fixed warning on x64

+++ trunk/test/codecs/HTMLEntityCodecTest.cpp
@@ -299,7 +299,7 @@ void DoWorkerThreadStuff()
   // *** Worker Threads ***
   for(unsigned int i=0; i<THREAD_COUNT; i++)
     {
-      int ret = pthread_create(&threads[i], NULL, WorkerThreadProc, (void*)i);
+      int ret = pthread_create(&threads[i], NULL, WorkerThreadProc, (void*)(intptr_t)i);
       if(0 != ret /*success*/)
         {
           BOOST_ERROR( "pthread_create failed (thread " << i << L"): " << strerror(errno) );
